% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2sf.R
\name{ah2sf}
\alias{ah2sf}
\title{Convert alpha hulls into sf POLYGON object}
\usage{
ah2sf(hull)
}
\arguments{
\item{hull}{Alpha hull}
}
\value{
a sf POLYGON object
}
\description{
Convert alpha hulls into sf POLYGON object
}
\examples{
library(alphahull)
set.seed(123)
x <- matrix(runif(100), nc = 2)
ahull_02 <- ahull(x, alpha = 0.2)
ahull_line_02 <- ah2sf(ahull_02)
}
