% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2sf.R
\name{ahull2lines}
\alias{ahull2lines}
\title{Convert alpha hulls into sf MULTILINESTRING object}
\usage{
ahull2lines(hull)
}
\arguments{
\item{hull}{Alpha hull}
}
\value{
a sf MULTILINESTRING object
}
\description{
Convert alpha hulls into sf MULTILINESTRING object
}
\examples{
library(alphahull)
set.seed(123)
x <- matrix(runif(100), nc = 2)
ahull_02 <- ahull(x, alpha = 0.2)
ahull_line_02 <- ahull2lines(ahull_02)
}
