% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcppahull2ahull.R
\name{rcppah2ah}
\alias{rcppah2ah}
\title{Fix alpha hulls generated from the package RcppAlphahul}
\usage{
rcppah2ah(x, orig_points, tol = 1e-10)
}
\arguments{
\item{x}{`ahull` object generated by the function ahull from the package RcppAlhahull}

\item{orig_points}{a matrix or data.frame of the original data points}
}
\value{
`ahull` object
}
\description{
Function written by chatGPT to fix improper `ahull` object generated by RcppAlphahull
}
