% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull2sf.R
\name{ah2sf}
\alias{ah2sf}
\title{Convert alpha hulls into sf POLYGON object}
\usage{
ah2sf(x, increment = 360, rnd = 10, crs = 4326, tol = 1e-04)
}
\arguments{
\item{x}{Alpha hull}
}
\value{
a sf POLYGON object
}
\description{
Function written by Andrew Bevan, found on R-sig-Geo, and modified by Pascal Title
modified to support sf objects 17 Nov 2022
added a fix to avoid error when casting from lines to polygons
}
\examples{
library(alphahull)
set.seed(123)
x <- matrix(runif(100), nc = 2)
ahull_02 <- ahull(x, alpha = 0.2)
ahull_line_02 <- ah2sf(ahull_02)
}
