% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_alpha_hulls.R
\name{make_alpha_hulls}
\alias{make_alpha_hulls}
\title{Estimating the species distribution range with alpha-hulls}
\usage{
make_alpha_hulls(loc_data, output_dir = NULL, binom_col = NULL,
  coordHeaders = NULL, fraction = 0.95, partCount = 10,
  initialAlpha = 2, alphaIncrement = 1, alphaDecrement = 1,
  maxIter = 2, default_buffer = 2000, other_buffers = c(2000),
  clipToCoast = "terrestrial", land_file = NULL,
  proj = "+proj=longlat +datum=WGS84", save.outputs = TRUE,
  do.parallel = TRUE, ncores = NULL, verbose = TRUE)
}
\arguments{
\item{loc_data}{A two-column matrix or data.frame or data.table or a character string specifying the path to your occurrence files directory (i.e. where species occurrence data files are stored) OR path to your species occurrence file (must be a csv file)}

\item{output_dir}{An optional character string specifying the path to the directory where to save the shapefiles generated.}

\item{binom_col}{A numeric value indicating which column number contains the binomials for the species (i.e. Genus species).Ignored if loc_data is not a csv file.}

\item{coordHeaders}{A character string vector of length 2 specifying the names of the columns indicating the longitude and latitude respectively.}

\item{fraction}{A numeric between 0 and 1 specifying the minimum fraction of occurrences that must be included in the polygon.}

\item{partCount}{A numeric integer specifying the maximum number of disjunct polygons that are allowed.}

\item{initialAlpha}{A numeric specifying the starting value for the parameter alpha.}

\item{alphaIncrement}{A numeric specifying the amount to increase alpha at each iteration}

\item{alphaDecrement}{A numeric specifying the amount to decrease alpha if the function fails after the first iterations}

\item{maxIter}{A numeric integer specifying the maximum number of iterations before trying to build circular buffers.}

\item{default_buffer}{A numeric specifying the buffering distance in meters around each point in case the buffer method equals 0. Default is 200 km.}

\item{other_buffers}{A numeric value or a vector of values of length \code{maxIter} specifying alternative buffer sizes (in meters) if the function fails after the first iterations.}

\item{clipToCoast}{A logical. Should the terrestrial \code{clipToCoast='terrestrial'}, the aquatic part \code{clipToCoast='aquatic'} or none part \code{clipToCoast='no'} part of the range be kept ?}

\item{land_file}{An optional character string specifying a path to a land basemap (shapefile or rds file) for clipping species polygon ranges by lands.}

\item{proj}{A character string specifying the projection of the coordinates}
}
\value{
None
}
\description{
Make an alpha-hull from the spatial distribution of species occurrence records
}
